Add client messenger

+++ client/client_messages.cpp
@@ -4,7 +4,7 @@
 
 std::ostream &operator<<(std::ostream &os, ClientMessage const &clientMessage) {
     return os << "session_id=" << clientMessage.session_id
-              << "    turn_direction=" << clientMessage.turn_direction
+              << "    turn_direction=" << static_cast<uint_fast32_t>(clientMessage.turn_direction)
               << "    next_expected_event_no=" << clientMessage.next_expected_event_no
               << "    player_name=" << clientMessage.player_name;
 }
